refactor: extract duplicated alpha-to-USD conversion and price hook to shared utils
