+++ src/components/issues/IssuesList.tsx
@@ -20,7 +20,8 @@ import {
 } from '@mui/material';
 import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import { IssueBounty } from '../../api/models/Issues';
-import { useStats } from '../../api';
+import { formatAlphaToUsd } from '../../utils/format';
+import { usePrices } from '../../hooks/usePrices';
 import { formatTokenAmount, formatDate } from '../../utils/format';
 import { getIssueStatusMeta } from '../../utils/issueStatus';
 import { STATUS_COLORS, TEXT_OPACITY } from '../../theme';
@@ -63,20 +64,7 @@ const IssuesList: React.FC<IssuesListProps> = ({
   const theme = useTheme();
   const [sortKey, setSortKey] = useState<SortKey>('id');
   const [sortDirection, setSortDirection] = useState<SortDirection>('desc');
-  const { data: dashStats } = useStats();
-  const taoPrice = dashStats?.prices?.tao?.data?.price ?? 0;
-  const alphaPrice = dashStats?.prices?.alpha?.data?.price ?? 0;
-
-  const toUsd = useCallback(
-    (alphaAmount: string): string | null => {
-      if (taoPrice <= 0 || alphaPrice <= 0) return null;
-      const amount = parseFloat(alphaAmount);
-      if (isNaN(amount) || amount === 0) return null;
-      const usd = amount * alphaPrice * taoPrice;
-      return `~${usd.toLocaleString('en-US', { style: 'currency', currency: 'USD', maximumFractionDigits: 0 })}`;
-    },
-    [alphaPrice, taoPrice],
-  );
+  const { taoPrice, alphaPrice } = usePrices();
   const headerCellSx = useMemo(
     () => ({
       fontFamily: '"JetBrains Mono", monospace',
@@ -371,7 +359,7 @@ const IssuesList: React.FC<IssuesListProps> = ({
           <TableBody>
             {sortedIssues.map((issue) => {
               const statusBadge = getIssueStatusMeta(issue.status);
-              const usdDisplay = toUsd(issue.targetBounty);
+              const usdDisplay = formatAlphaToUsd(issue.targetBounty, taoPrice, alphaPrice);
 
               return (
                 <TableRow
